Update BAM Encoding.py
Update the counting matrix's dimension (did a reverse of the previous counting matrix since we need to map to the correct dimension 12xL, the previous version is wrong since it generates a matrix dimension Lx12). Also the position used inside the matrix should be the local position inside the current region instead of the global position in the whole genome. So we use (pileupcolumn.pos - start) instead of (pileupcolumn.pos).